feat: score_delta를 composite_score 기준으로 변경
sentiment_score(정수) 대신 composite_score(float) 기준으로 전일 대비 delta 계산.
어제 파일에 composite_score가 없으면 sentiment_score로 폴백.
score_delta 타입을 int → float으로 변경(Pydantic schema, TS types 동기화).

+++ backend/api/schemas.py
@@ -179,7 +179,7 @@ class SymbolSentiment(BaseModel):
     bot_suspected: str
     confidence: str
     source: str
-    score_delta: Optional[int] = None
+    score_delta: Optional[float] = None
     intraday_shift: Optional[str] = None
 
 class MarketSentiment(BaseModel):
